FogBugz #180526: Add USB dual-role support
This commit enables the DWC2 USB driver to be a dual-role OTG driver.
This means that the driver can be in host or peripheral mode when the appropriate
connector is used. When an A-cable is plugged in, the driver behaves in host
mode, and when a B-cable is used, the driver will be in peripheral mode.

This commit:
- Unifies the dwc2 and s3c-hsotg driver into a single driver.
- Deletes the defines in s3c-hsotg.
- Adds a aingle IRQ handler for both drivers.

Signed-off-by: Dinh Nguyen <dinguyen@altera.com>