FogBugz #187411: Add TSE MSGDMA and SGMDA Hybrid Driver
This commit adds the TSE MSGDMA and SGDMA hybrid driver. The
driver checks the device tree bindings compatibility string
to determine what mode the driver is operating in, and
probes the device tree for required resources accordingly.
SGDMA will be deprecated very soon by Altera, but is still
supported in this driver for customers that still use SGDMA.
The driver serves as a good example of how to use two
different types of soft DMA IP. The driver was tested on
ARM and NIOS systems, and issues specific to those systems
are handled by the driver such as misaligned exceptions on
the NIOS processor.

V3:
- Remove maintainer changes
- Correct git commit log line lengths

V2:
- Update per Crucible review 596
- Update bindings descriptions altera_tse.txt per review comments
- Changed altera_tse_driver to alteratse in Makefile and Kconfig
- Added entry in the Maintainers file
- changed altera_sgdmahw.h to use BIT macro instead of like (1<<0)
- corrected copyrights per review comments
- moved cacheflush.h to bottom of header inclusions in altera_tse.c
- use apis to extract lower and upper 32 bits of dma addresss
- remove barriers since dma apis ensure memory consistency
- add max-frame-size property
- change platform id from AVALON to "platform"
- use of function of_get_phy_mode instead of custom function
- remove use of netif_carrier_off, not needed in driver
- move probing of phy from open to the probe function

Signed-off-by: Vince Bridgers <vbridger@altera.com>