FogBugz #205319: nios2: Use initi and initd to initialize caches
Change flushi to initi and flushd to initd instructions.

According to nios2 handbook:
- Use initi instruction initializes a single
instruction cache line. Do not use the flushi instruction because
it might cause undesired effects when used to initialize the
instruction cache in future Nios II implementations.
- Use initd instruction initializes a single data cache line. Do not
use the flushd instruction for this purpose, because it writes dirty
lines back to memory.

[1]http://www.altera.com/literature/hb/nios2/n2sw_nii52007.pdf

Signed-off-by: Ley Foon Tan <lftan@altera.com>