FogBugz #208682: Denali NAND driver misidentifies Hynix chips
The denali driver was not reading enough id bytes to determine the
parameters for a Hynix chip.  The chip was therefore not identified
properly, and the page size and oob size were calculated incorrectly.

This resulted in a message stating the OOB area was not big enough
for 8-bit ECC.

Signed-off-by: Graham Moore <grmoore@altera.com>